+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - GP2-1168 - export plan lessons
 - NOTICKET - upgrade admin include export plan data
 - GP2-1139 - save EP progress
+- GP2-1382 - getting paid structure be
 
 ### Fixed bugs
 - GP2-1375 - Fix image-types for advice and markets hero